isBeingTranspiled
